+++ authority-portal-frontend/src/app/pages/participant-own-connector-detail-page/participant-own-connector-detail-page/participant-own-connector-detail-page.component.ts
@@ -22,13 +22,13 @@ import {
 } from 'rxjs';
 import {Store} from '@ngxs/store';
 import {GlobalStateUtils} from 'src/app/core/global-state/global-state-utils';
-import {
-  getConnectorStatusText,
-  getConnectorsTypeClasses,
-  getConnectorsTypeText,
-} from 'src/app/core/utils/ui-utils';
+import {getConnectorsTypeClasses} from 'src/app/core/utils/ui-utils';
 import {TitleBarConfig} from 'src/app/shared/common/portal-layout/title-bar/title-bar.model';
 import {ChildComponentInput} from 'src/app/shared/common/slide-over/slide-over.model';
+import {
+  getConnectorStatusText,
+  getConnectorsTypeText,
+} from '../../../core/utils/mappers/dto-ui-mapper';
 import {DeleteOwnConnector} from '../../participant-own-connector-list-page/state/participant-own-connector-list-page-actions';
 import {
   RefreshConnector,
