+++ authority-portal-frontend/src/app/pages/sp-connector-detail-page/sp-connector-detail-page/sp-connector-detail-page.component.ts
@@ -14,12 +14,10 @@ import {Component, Inject, OnDestroy, OnInit} from '@angular/core';
 import {Subject, takeUntil} from 'rxjs';
 import {Store} from '@ngxs/store';
 import {ConnectorDetailsDto} from '@sovity.de/authority-portal-client';
-import {
-  getConnectorsTypeClasses,
-  getConnectorsTypeText,
-} from 'src/app/core/utils/ui-utils';
+import {getConnectorsTypeClasses} from 'src/app/core/utils/ui-utils';
 import {TitleBarConfig} from 'src/app/shared/common/portal-layout/title-bar/title-bar.model';
 import {ChildComponentInput} from 'src/app/shared/common/slide-over/slide-over.model';
+import {getConnectorsTypeText} from '../../../core/utils/mappers/dto-ui-mapper';
 import {DeleteProvidedConnector} from '../../sp-connector-list-page/state/sp-connector-list-page-actions';
 import {
   RefreshConnector,
