+++ authority-portal-frontend/src/app/shared/business/shared-connector-detail/shared-connector-detail.component.ts
@@ -13,7 +13,7 @@
 import {Component, HostBinding, Input} from '@angular/core';
 import {ConnectorDetailsDto} from '@sovity.de/authority-portal-client';
 import {ClipboardUtils} from 'src/app/core/utils/clipboard-utils';
-import {getConnectorStatusText} from 'src/app/core/utils/ui-utils';
+import {getConnectorStatusText} from '../../../core/utils/mappers/dto-ui-mapper';
 
 @Component({
   selector: 'app-shared-connector-detail',
